move patches to proper folders
