+++ SolastaCommunityExpansion/Patches/SrdAndHouseRules/BestowCurseNoConcentrationRequired/CharacterActionCastSpellPatcher.cs
@@ -1,7 +1,7 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using HarmonyLib;
 
-namespace SolastaCommunityExpansion.Patches.Bugfix
+namespace SolastaCommunityExpansion.Patches.SrdAndHouseRules.BestowCurseNoConcentrationRequired
 {
     [HarmonyPatch(typeof(CharacterActionCastSpell), "StartConcentrationAsNeeded")]
     [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
